refactor: extract host/path utility helpers into transfer-host-utils module
